refactor(domain): Lotto와 LottoTickets 클래스를 domain 패키지로 이동
- 프로젝트 구조 개선하기 위해 클래스 이동
