+++ src/main/kotlin/lotto/controller/LottoController.kt
@@ -1,6 +1,6 @@
 package lotto.controller
 
-import lotto.Lotto
+import lotto.domain.Lotto
 import lotto.view.InputView
 import lotto.view.OutputView
