+++ src/main/kotlin/lotto/domain/LottoTickets.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class LottoTickets(val tickets: List<Lotto>) {
     fun size(): Int = tickets.size
